backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
requests==2.32.3
python-dotenv==1.0.1
reportlab==4.2.0
